+++ api/src/main/resources/application-local.properties
@@ -95,7 +95,8 @@ launchDarkly.flushInterval=10
 
 bulk.scanning.payments.processed.url=http://localhost:8004
 
-iac.supplementary.info.url=http://localhost:8092
+#iac.supplementary.info.url=http://localhost:8092
+iac.supplementary.info.url=http://ia-case-access-api-aat.service.core-compute-aat.internal
 
 idam.s2s-auth.totp_secret=${PAYMENT_S2S_SECRET:AAAAAAAAAAAAAAAA}
 idam.s2s-auth.microservice=${PAYMENT_S2S_MICROSERVICE_NAME:internal}
