--- model/src/main/java/uk/gov/hmcts/payment/api/service/IACService.java
@@ -1,10 +1,0 @@
-package uk.gov.hmcts.payment.api.service;
-
-import org.springframework.http.ResponseEntity;
-import uk.gov.hmcts.payment.api.dto.SupplementaryDetailsResponse;
-
-import java.util.List;
-
-public interface IACService {
-    ResponseEntity<SupplementaryDetailsResponse> getIacSupplementaryInfo(List<String> iacCcdCaseNos,String serviceToken);
-}
